Add MINUS_DI, MINUS_DM, SAR, and MAVP functions

+++ tabox/ta_func/ta_utility.py
@@ -11,126 +11,126 @@ from ..retcode import TA_RetCode
 PI = 3.14159265358979323846
 
 
-def TA_REAL_EQ(x: cython.float, v: cython.float, ep: cython.float):
+def TA_REAL_EQ(x: cython.double, v: cython.double, ep: cython.double):
     return ((v - ep) < x) and (x < (v + ep))
 
 
-def TA_IS_ZERO(v: cython.float):
+def TA_IS_ZERO(v: cython.double):
     return ((-0.00000001) < v) and (v < 0.00000001)
 
 
-def TA_IS_ZERO_OR_NEG(v: cython.float):
+def TA_IS_ZERO_OR_NEG(v: cython.double):
     return v < 0.00000001
 
 
 # 数学函数
-def std_floor(x: cython.float) -> cython.float:
+def std_floor(x: cython.double) -> cython.double:
     return math.floor(x)
 
 
-def std_ceil(x: cython.float) -> cython.float:
+def std_ceil(x: cython.double) -> cython.double:
     return math.ceil(x)
 
 
-def std_fabs(x: cython.float) -> cython.float:
+def std_fabs(x: cython.double) -> cython.double:
     return abs(x)
 
 
-def std_atan(x: cython.float) -> cython.float:
+def std_atan(x: cython.double) -> cython.double:
     return math.atan(x)
 
 
-def std_cos(x: cython.float) -> cython.float:
+def std_cos(x: cython.double) -> cython.double:
     return math.cos(x)
 
 
-def std_sin(x: cython.float) -> cython.float:
+def std_sin(x: cython.double) -> cython.double:
     return math.sin(x)
 
 
-def std_sqrt(x: cython.float) -> cython.float:
+def std_sqrt(x: cython.double) -> cython.double:
     return math.sqrt(x)
 
 
-def std_tanh(x: cython.float) -> cython.float:
+def std_tanh(x: cython.double) -> cython.double:
     return math.tanh(x)
 
 
-def std_tan(x: cython.float) -> cython.float:
+def std_tan(x: cython.double) -> cython.double:
     return math.tan(x)
 
 
-def std_sinh(x: cython.float) -> cython.float:
+def std_sinh(x: cython.double) -> cython.double:
     return math.sinh(x)
 
 
-def std_log10(x: cython.float) -> cython.float:
+def std_log10(x: cython.double) -> cython.double:
     return math.log10(x)
 
 
-def std_log(x: cython.float) -> cython.float:
+def std_log(x: cython.double) -> cython.double:
     return math.log(x)
 
 
-def std_exp(x: cython.float) -> cython.float:
+def std_exp(x: cython.double) -> cython.double:
     return math.exp(x)
 
 
-def std_cosh(x: cython.float) -> cython.float:
+def std_cosh(x: cython.double) -> cython.double:
     return math.cosh(x)
 
 
-def std_asin(x: cython.float) -> cython.float:
+def std_asin(x: cython.double) -> cython.double:
     return math.asin(x)
 
 
-def std_acos(x: cython.float) -> cython.float:
+def std_acos(x: cython.double) -> cython.double:
     return math.acos(x)
 
 
 # Utility Functions
-def round_pos(x: cython.float) -> cython.float:
+def round_pos(x: cython.double) -> cython.double:
     return std_floor(x + 0.5)
 
 
-def round_neg(x: cython.float) -> cython.float:
+def round_neg(x: cython.double) -> cython.double:
     return std_ceil(x - 0.5)
 
 
-def round_pos_2(x: cython.float) -> cython.float:
+def round_pos_2(x: cython.double) -> cython.double:
     return (std_floor((x * 100.0) + 0.5)) / 100.0
 
 
-def round_neg_2(x: cython.float) -> cython.float:
+def round_neg_2(x: cython.double) -> cython.double:
     return (std_ceil((x * 100.0) - 0.5)) / 100.0
 
 
-def PER_TO_K(per: cython.int) -> cython.float:
+def PER_TO_K(per: cython.int) -> cython.double:
     return 2.0 / (per + 1)
 
 
 # K-Line Related Functions
-def TA_REALBODY(inClose: cython.float, inOpen: cython.float) -> cython.float:
+def TA_REALBODY(inClose: cython.double, inOpen: cython.double) -> cython.double:
     return std_fabs(inClose - inOpen)
 
 
 def TA_UPPERSHADOW(
-    inHigh: cython.float, inClose: cython.float, inOpen: cython.float
-) -> cython.float:
+    inHigh: cython.double, inClose: cython.double, inOpen: cython.double
+) -> cython.double:
     return inHigh - (inClose if inClose >= inOpen else inOpen)
 
 
 def TA_LOWERSHADOW(
-    inClose: cython.float, inOpen: cython.float, inLow: cython.float
-) -> cython.float:
+    inClose: cython.double, inOpen: cython.double, inLow: cython.double
+) -> cython.double:
     return (inOpen if inClose >= inOpen else inClose) - inLow
 
 
-def TA_HIGHLOWRANGE(inHigh: cython.float, inLow: cython.float) -> cython.float:
+def TA_HIGHLOWRANGE(inHigh: cython.double, inLow: cython.double) -> cython.double:
     return inHigh - inLow
 
 
-def TA_CANDLECOLOR(inClose: cython.float, inOpen: cython.float) -> cython.int:
+def TA_CANDLECOLOR(inClose: cython.double, inOpen: cython.double) -> cython.int:
     return 1 if inClose >= inOpen else -1
